Second part of major reworking of the VAMM system, to fix bugs related to determining erased pages. Changed project version number to V3.0. Need to fix the returning of the stored data sizes and V3.0 will be ready for release.

+++ StorageManager.h
@@ -30,7 +30,7 @@
 	
 	#define SM_MAX_FUSELOCKBITS     20
 	
-	#define SM_EEPROM_OFFSET        ((DF_DATAFLASH_PAGES / 2) * DF_INTERNALDF_BUFFBYTES)
+	#define SM_EEPROM_OFFSET        (((unsigned long)DF_DATAFLASH_PAGES / 2) * DF_INTERNALDF_BUFFBYTES)
 	
 	// EXTERNAL VARIABLES:
 	extern uint8_t  MemoryType;
